==> part_2_basic_cleaning.ipynb <==
Basic Data Cleaning

from the looks of it - only study results have demographic information. So we will only be looking at those.

example of a study result
https://clinicaltrials.gov/ct2/show/results/NCT03698591?cond=TMS&draw=3&rank=7

example of no study results
https://clinicaltrials.gov/ct2/show/results/NCT04982757?cond=TMS&draw=3&rank=8

disease = 'tdcs'

from os import system
import pandas as pd
import requests
import datetime as dt
# from pytrials.client import ClinicalTrials
import json
import ipywidgets as widgets
import numpy as np
np.random.seed(10031975)
import matplotlib.pyplot as plt
import numpy as np
from matplotlib import rcParams, cycler
import matplotlib.pyplot as plt
from matplotlib.ticker import AutoMinorLocator
# from openpyxl import Workbook
import openpyxl

# cleaning = pd.read_excel('data/{}_cleaning.xlsx'.format(disease))
cleaning = pd.read_excel('1-'+disease+'_starting_Data.xlsx')


# cleaning.head()
cleaning.shape

1454


# result.head()
i=-1
for col in cleaning.columns:
   i=i+1
   print(i, col)

#OverallStatus

Terminated ( remove)
Completed ( keep)
Recruiting ( keep)
Not yet recruiting ( remove)
Withdrawn ( remove)
Unknown status ( remove)
Active, not recruiting ( keep)
Suspended ( remove)
Enrolling by invitation ( keep )

First lets clean up on the status ( col 15 - overall status)
removing 
Suspended,Not yet recruiting, Withdrawn

# df_filtered = cleaning[cleaning['OverallStatus'] == 'Withdrawn' ]

# cleaning.drop(df.index[df['Col1'] == 0], inplace=True)
cleaning.drop(cleaning.index[cleaning['OverallStatus'] == ('Withdrawn')], inplace=True)

# cleaning.drop(cleaning.index[cleaning['OverallStatus'] == ('Terminated')], inplace=True)

cleaning.drop(cleaning.index[cleaning['OverallStatus'] == ('Not yet recruiting')], inplace=True)

cleaning.drop(cleaning.index[cleaning['OverallStatus'] == ('Suspended')], inplace=True)

cleaning.drop(cleaning.index[cleaning['OverallStatus'] == ('Unknown status')], inplace=True)


cleaning.shape

cleaning.head()

Now since we are looking at enrollment - lets remove all the studies that do not have at least one enrolled person


cleaning.drop(cleaning.index[cleaning['EnrollmentCount'] < 1 ], inplace=True)

# remove null values

cleaning = cleaning[cleaning['EnrollmentCount'].notna()]

# convert to int
cleaning = cleaning.astype({'EnrollmentCount': 'int32'})
# cleaning = cleaning[cleaning['EnrollmentCount'].notna()]


cleaning.shape

# for col in cleaning:
#     print(cleaning['EnrollmentCount'])

Now I need to remove Anticipated and other enrollment types


cleaning.shape

#  cleaning['EnrollmentType'].replace('', np.nan, inplace=True)



cleaning.drop(cleaning.index[cleaning['EnrollmentType'] == ('Anticipated')], inplace=True)

# this removes the null ones
cleaning.dropna(subset=['EnrollmentType'], inplace=True)

# verifyed on 
# https://clinicaltrials.gov/ct2/show/NCT00000748?cond=NCT00000748&draw=2&rank=1
# https://clinicaltrials.gov/ct2/show/NCT00006285?cond=NCT00006285&draw=2&rank=1
# https://clinicaltrials.gov/ct2/show/results/NCT00001783?cond=NCT00001783&draw=2&rank=1
# https://clinicaltrials.gov/ct2/show/results/NCT00054652?cond=NCT00054652&draw=2&rank=1


cleaning.shape

cleaned = cleaning

#reindex the dataframe

cleaned.reindex()
cleaned = cleaned.drop(cleaned.columns[[0]], axis=1)
cleaned.reset_index(drop=True, inplace=True)

cleaned.shape

cleaned

Verify random results


# cleaned['pre_link'] = pd.Series(['https://www.clinicaltrials.gov/ct2/show/results/' for x in range(len(cleaned.index))])
#

# cleaned["link"] = "=HYPERLINK(\"https://www.clinicaltrials.gov/ct2/show/results/"+cleaned["NCTId"].astype(str)+"\", \"+'Ric'.astype(str)+'\')"
cleaned["link"] = "=HYPERLINK(\"https://www.clinicaltrials.gov/ct2/show/results/"+cleaned["NCTId"].astype(str)+"\", \""+cleaned["NCTId"].astype(str)+"\")"
# cleaned["link"] = 
 # org
# cleaned["link"] = "=HYPERLINK(\"https://www.clinicaltrials.gov/ct2/show/results/"+cleaned["NCTId"].astype(str)+"\", \""+cleaned["NCTId"].astype(str)+"\")"

# cleaned = cleaned({'link':['=HYPERLINK("https://www.clinicaltrials.gov/ct2/show/"+cleaned["NCTId"].astype(str)+", "+cleaned["NCTId"].astype(str)+")']})




# cleaned = cleaned({'link':['=HYPERLINK("https://www.clinicaltrials.gov/ct2/show/"+cleaned["NCTId"].astype(str)+", "+cleaned["NCTId"].astype(str)+")']})

cleaned['link']

cleaned.shape

fixed some data types

# cleaned['StudyFirstSubmitDate']= pd.to_datetime(cleaned['StudyFirstSubmitDate']).dt.date
# cleaned['LastUpdateSubmitDate']= pd.to_datetime(cleaned['LastUpdateSubmitDate']).dt.date
# cleaned['CompletionDate']= pd.to_datetime(cleaned['CompletionDate']).dt.date


cleaned['StudyFirstSubmitDate']= pd.to_datetime(cleaned['StudyFirstSubmitDate'])
cleaned['LastUpdateSubmitDate']= pd.to_datetime(cleaned['LastUpdateSubmitDate'])
cleaned['CompletionDate']= pd.to_datetime(cleaned['CompletionDate'])


cleaned.dtypes

# cleaned = cleaned.drop(cleaning['pre_link'], axis=1)
# cleaned = cleaned.drop('pre_link', axis=1)
# cleaned = cleaned.drop(cleaned.columns[[0]], axis=1)

cleaned.dtypes

EXCEL IS NOT THE BEST - DONT USE in PRODUCTION

# precleaning


from datetime import datetime
import xlsxwriter

options = {}
options['strings_to_formulas'] = False
options['strings_to_urls'] = False
options['strings_to_numbers'] = False
options['strings_to_datetime'] = False
options['strings_to_columns'] = False
options['strings_to_comments'] = False
options['strings_to_format'] = False
options[ datetime] = 'yyyy/mm/dd'

# options['strings_to_urls'] = False
writer = pd.ExcelWriter(('2-'+disease+'_cleaned_Data.xlsx'),engine='xlsxwriter',options=options)
cleaned.to_excel(writer, 'Cleaned', index=False)
writer.save()

cleaned.to_csv('2-'+disease+'_cleaned_use_me.csv', index = False)

cleaned.shape

# # result.to_csv(r'data_raw.csv', index = False)
# cleaned.to_excel(disease+'_2.xlsx', index = False)
