==> part_3_flat_file.ipynb <==
disease = 'tdcs'

from os import system
import pandas as pd
import requests
import datetime as dt
# from pytrials.client import ClinicalTrials
import json
import ipywidgets as widgets
import numpy as np
np.random.seed(10031975)
import matplotlib.pyplot as plt
import numpy as np
from matplotlib import rcParams, cycler
import matplotlib.pyplot as plt
from matplotlib.ticker import AutoMinorLocator
# from openpyxl import Workbook
import openpyxl

clean2 = pd.read_csv('2-'+disease+'_cleaned_use_me.csv')

baseline = pd.read_csv('baseline_measurements.txt', sep='|', low_memory=False
# index_col=False,  
    # dtype={"param_value": int, "dispersion_value": str}  
    )
                #  names=['protocol', 'server', 'type', 'value'])

10 - 10 param_value
13- 13 dispersion_value

# result.head()
i=-1
for col in baseline.columns:
   i=i+1
   print(i, col)

baseline.info()

baseline.rename(columns = {'nct_id':'NCTId'}, inplace = True)
baseline = baseline.drop('id', axis=1)
baseline = baseline.drop('explanation_of_na', axis=1)
#used for stats - we don't need it
baseline = baseline.drop('dispersion_type', axis=1)
baseline = baseline.drop('dispersion_value', axis=1)
baseline = baseline.drop('dispersion_value_num', axis=1)
baseline = baseline.drop('dispersion_lower_limit', axis=1)
baseline = baseline.drop('dispersion_upper_limit', axis=1)
# baseline = baseline.drop('dispersion_value', axis=1)

classification is important 
category is important
Title is important (first search on this )


baseline.head()
baseline.shape

baseline.dtypes

baseline[["param_value", "param_value_num"]] = baseline[["param_value", "param_value_num"]].apply(pd.to_numeric, errors='coerce') 

# a = baseline['dispersion_type'].unique()
# print((a))


# clean2.shape
# clean2.head()

# baseline.shape
# baseline.head()

result = pd.merge(clean2, baseline,  on='NCTId', how='left')

result.shape

result.to_csv('3-'+disease+'_results.csv', index = False)
